graph : move KV cache build functions to llama_context impl
ggml-ci
